Move session out of our module
